Refactor and add onSucces to auth request
