+++ src/network/firebaseAuth/firebaseAuth.ts
@@ -2,35 +2,24 @@ import {
 	createUserWithEmailAndPassword,
 	signInWithEmailAndPassword,
 } from 'firebase/auth';
-import { auth } from '../libs/firebase/firebase.config';
-
-type AuthType = 'signup' | 'login';
-
-export interface FirebaseAuthProps {
-	authType: AuthType;
-	values: {
-		email: string;
-		password: string;
-	};
-	onSucces?: () => void;
-	onError?: (message: string) => void;
-}
+import { auth } from '../../libs/firebase/firebase.config';
+import { FirebaseAuthProps } from './firebaseAuthInterfaces';
+import { PATHS } from '../../router/paths';
 
 export const firebaseAuth = ({
 	authType,
 	values,
 	onError,
+	onSuccess,
 }: FirebaseAuthProps) => {
 	const { email, password } = values;
-	console.log(authType);
-
 	const authFunction = {
 		login: signInWithEmailAndPassword,
 		signup: createUserWithEmailAndPassword,
 	};
 	authFunction[authType](auth, email, password)
-		.then((userCredential) => {
-			console.log(userCredential);
+		.then(() => {
+			onSuccess?.(PATHS.ROOT);
 		})
 		.catch((error: string) => {
 			onError?.(error);
