+++ src/network/firebaseAuth/firebaseAuthInterfaces.ts
@@ -1,0 +1,11 @@
+type AuthType = 'signup' | 'login';
+
+export interface FirebaseAuthProps {
+	authType: AuthType;
+	values: {
+		email: string;
+		password: string;
+	};
+	onSuccess?: (route: string) => void;
+	onError?: (message: string) => void;
+}
